requirements.txt
Flask==2.0.1
requests==2.26.0
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
pytest==6.2.4
beautifulsoup4==4.9.3
lxml==4.6.3
SQLAlchemy==1.4.22
python-dotenv==0.19.0
gunicorn==20.1.0
PyJWT==2.1.0
Flask-JWT-Extended==4.4.4
